Cast BYTE24.. to uint8_t

+++ _treufunk.h
@@ -40,9 +40,9 @@
 /**
  * Macros for H, M and L byte of 24 bit value
  */
-#define BIT24_H_BYTE(c) (((c) & 0xFF0000) >> 16)
-#define BIT24_M_BYTE(c) (((c) & 0x00FF00) >> 8)
-#define BIT24_L_BYTE(c) ((c) & 0x0000FF)
+#define BIT24_H_BYTE(c) ((uint8_t)(((c) & 0xFF0000) >> 16))
+#define BIT24_M_BYTE(c) ((uint8_t)(((c) & 0x00FF00) >> 8))
+#define BIT24_L_BYTE(c) ((uint8_t)((c) & 0x0000FF))
 
 
 #ifdef __cplusplus
